+++ AdminClient/src/main/java/org/familydirectory/sdk/adminclient/events/delete/DeleteEvent.java
@@ -2,7 +2,6 @@ package org.familydirectory.sdk.adminclient.events.delete;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Scanner;
 import java.util.function.Predicate;
 import org.familydirectory.assets.ddb.enums.DdbTable;
@@ -14,29 +13,14 @@ import org.familydirectory.sdk.adminclient.events.model.EventHelper;
 import org.familydirectory.sdk.adminclient.utility.MemberPicker;
 import org.jetbrains.annotations.NotNull;
 import software.amazon.awssdk.services.cognitoidentityprovider.CognitoIdentityProviderClient;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.AdminDeleteUserRequest;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.AttributeType;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.ListUserPoolsRequest;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.ListUsersRequest;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.ListUsersResponse;
-import software.amazon.awssdk.services.cognitoidentityprovider.model.UserType;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 import software.amazon.awssdk.services.dynamodb.model.Delete;
-import software.amazon.awssdk.services.dynamodb.model.DeleteItemRequest;
 import software.amazon.awssdk.services.dynamodb.model.QueryRequest;
 import software.amazon.awssdk.services.dynamodb.model.QueryResponse;
 import software.amazon.awssdk.services.dynamodb.model.TransactWriteItem;
 import software.amazon.awssdk.services.dynamodb.model.TransactWriteItemsRequest;
 import software.amazon.awssdk.services.dynamodb.model.Update;
-import software.amazon.awssdk.services.sesv2.SesV2Client;
-import software.amazon.awssdk.services.sesv2.model.Body;
-import software.amazon.awssdk.services.sesv2.model.Content;
-import software.amazon.awssdk.services.sesv2.model.Destination;
-import software.amazon.awssdk.services.sesv2.model.EmailContent;
-import software.amazon.awssdk.services.sesv2.model.Message;
-import software.amazon.awssdk.services.sesv2.model.SendEmailRequest;
-import static java.lang.System.getenv;
 import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static java.util.Objects.requireNonNull;
@@ -175,82 +159,8 @@ class DeleteEvent implements EventHelper {
                                                                                     .get(CognitoTableParameter.ID.jsonFieldName())).map(AttributeValue::s)
                                                                                                                                    .filter(Predicate.not(String::isBlank))
                                                                                                                                    .orElseThrow();
-
-            this.dynamoDbClient.deleteItem(DeleteItemRequest.builder()
-                                                            .tableName(DdbTable.COGNITO.name())
-                                                            .key(singletonMap(CognitoTableParameter.ID.jsonFieldName(), AttributeValue.fromS(ddbMemberCognitoSub)))
-                                                            .build());
-
-            final String userPoolId = this.getUserPoolId();
-
-            final ListUsersRequest listUsersRequest = ListUsersRequest.builder()
-                                                                      .filter("sub = \"%s\"".formatted(ddbMemberCognitoSub))
-                                                                      .limit(1)
-                                                                      .userPoolId(userPoolId)
-                                                                      .build();
-            final UserType ddbMemberCognitoUser = ofNullable(this.cognitoClient.listUsers(listUsersRequest)).map(ListUsersResponse::users)
-                                                                                                            .filter(Predicate.not(List::isEmpty))
-                                                                                                            .map(List::getFirst)
-                                                                                                            .orElseThrow();
-            final String ddbMemberCognitoUsername = Optional.of(ddbMemberCognitoUser)
-                                                            .map(UserType::username)
-                                                            .filter(Predicate.not(String::isBlank))
-                                                            .orElseThrow();
-            this.cognitoClient.adminDeleteUser(AdminDeleteUserRequest.builder()
-                                                                     .userPoolId(userPoolId)
-                                                                     .username(ddbMemberCognitoUsername)
-                                                                     .build());
-
-            final String ddbMemberCognitoEmail = Optional.of(ddbMemberCognitoUser)
-                                                         .map(UserType::attributes)
-                                                         .filter(Predicate.not(List::isEmpty))
-                                                         .orElseThrow()
-                                                         .stream()
-                                                         .filter(attr -> attr.name()
-                                                                             .equalsIgnoreCase("email"))
-                                                         .findFirst()
-                                                         .map(AttributeType::value)
-                                                         .filter(s -> s.contains("@"))
-                                                         .orElseThrow();
-            sendDeletionNoticeEmail(singletonList(ddbMemberCognitoEmail));
+            this.deleteCognitoAccountAndNotify(this.cognitoClient, ddbMemberCognitoSub);
         }
-    }
-
-    public static
-    void sendDeletionNoticeEmail (final @NotNull List<String> addresses) {
-        final Message message = Message.builder()
-                                       .subject(Content.builder()
-                                                       .data("Notice of Account Deletion")
-                                                       .build())
-                                       .body(Body.builder()
-                                                 .text(Content.builder()
-                                                              .data("Your account has been irreversibly deleted.")
-                                                              .build())
-                                                 .build())
-                                       .build();
-        try (final SesV2Client sesClient = SesV2Client.create()) {
-            sesClient.sendEmail(SendEmailRequest.builder()
-                                                .destination(Destination.builder()
-                                                                        .toAddresses(addresses)
-                                                                        .build())
-                                                .content(EmailContent.builder()
-                                                                     .simple(message)
-                                                                     .build())
-                                                .fromEmailAddress("no-reply@%s".formatted(requireNonNull(getenv("ORG_FAMILYDIRECTORY_HOSTED_ZONE_NAME"))))
-                                                .build());
-        }
-    }
-
-    @NotNull
-    private
-    String getUserPoolId () {
-        return ofNullable(this.cognitoClient.listUserPools(ListUserPoolsRequest.builder()
-                                                                               .maxResults(1)
-                                                                               .build())
-                                            .userPools()).filter(Predicate.not(List::isEmpty))
-                                                         .map(l -> l.getFirst()
-                                                                    .id())
-                                                         .orElseThrow();
     }
 
     @Override
